backend/requirements.txt
fastapi
uvicorn
imagehash
exifread
Pillow
pydantic
Requests
python-multipart
